--- Spring/src/main/java/GFO/Spring/domain/email/presentation/controller/ApiResponses.java
@@ -1,1 +1,0 @@
-package GFO.Spring.domain.email.presentation.controller;
