+++ Spring/src/main/java/GFO/Spring/domain/email/presentation/controller/EmailController.java
@@ -1,15 +1,11 @@
 package GFO.Spring.domain.email.presentation.controller;
 
-import GFO.Spring.domain.email.service.EmailService;
+import GFO.Spring.domain.email.service.MailSendService;
 import GFO.Spring.domain.user.presentation.dto.UserDto;
+import GFO.Spring.domain.user.presentation.dto.request.SignupRequest;
 import GFO.Spring.domain.user.service.UserService;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -20,20 +16,19 @@ public class EmailController {
     @Autowired
     private UserService userService;
     @Autowired
-    private EmailService mss;
+    private MailSendService mss;    // mss 는MailSendService 라는 뜻!
 
     @PostMapping("/signup")
-    public void signUp(@RequestBody UserDto userDto){
+    public void signUp(@RequestBody SignupRequest SignupRequest){
         //DB에 기본 정보 삽입
-        userService.signup(userDto);
+        userService.signUp(SignupRequest);
 
-        //authKey 생성 & 이메일 발송
-        String authKey = mss.sendAuthMail(userDto.getUserEmail());
+        //authKey 생성  이메일 발송
+        String authKey = mss.sendAuthMail(SignupRequest.getEmail());
     }
 
-    @GetMapping("member/signUpConfirm")
+    @GetMapping("member/signup-confirm")
     public void signUpConfirm(@RequestParam String email){
-        //email, authKey 가 일치할경우 authStatus 업데이트
         userService.updateAuthStatus(email);
         System.out.println("email = " + email);
     }
