+++ Spring/src/main/java/GFO/Spring/domain/email/service/MailSendService.java
@@ -5,26 +5,22 @@ import GFO.Spring.domain.user.presentation.dto.UserDto;
 import GFO.Spring.domain.user.repository.UserRepository;
 import GFO.Spring.global.MailUtils;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.mail.MailException;
-import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.JavaMailSenderImpl;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.mail.Message.RecipientType;
 import javax.mail.MessagingException;
-import javax.mail.internet.MimeMessage;
 import java.io.UnsupportedEncodingException;
 import java.util.Random;
-import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
-public class MailSendService implements EmailService{
+public class MailSendService {
     private JavaMailSenderImpl mailSender;
     private UserRepository userRepository;
-    // 난수의 크기 지정 변수.
+    private PasswordEncoder passwordEncoder;
+    // 난수의 크기 변수
     private int size;
     //인증 키 생성
     private String getKey(int size){
@@ -32,44 +28,56 @@ public class MailSendService implements EmailService{
         return getAuthCode();
     }
 
-    //인증코드 난수 발생
+    //인증코드 발생
     private String getAuthCode() {
         Random random = new Random();
-        StringBuffer buffer = new StringBuffer();
-        int num = 0;
-
-        while (buffer.length() < size){
-            num = random.nextInt(10);
-            buffer.append(num);
-        }
-        return buffer.toString();
+        return String.valueOf(random.nextInt(888888) + 111111);
     }
 
     //인증메일 보내기
     public String sendAuthMail(String email) {
         //6자리 난수 인증번호 생성
-        String authKey = getKey(6);
+        String authKey = getAuthCode();
 
-        //인증메일 보내기
         try {
             MailUtils sendMail = new MailUtils(mailSender);
             sendMail.setSubject("회원가입 이메일 인증");
             sendMail.setText(new StringBuffer().append("<h1>[이메일 인증]</h1>")
                     .append("<p>아래 링크를 클릭하시면 이메일 인증이 완료됩니다.</p>")
-                    .append("<a href='http://localhost:8080/user/member/signUpConfirm?email=")
+                    .append("<a href='http://localhost:8080/user/signup-confirm?email=")
                     .append(email)
                     .append("' target='_blenk'>이메일 인증 확인</a>")
                     .toString());
             sendMail.setFrom("s22055@gsm.hs.kr" , "채종인");
             sendMail.setTo(email);
             sendMail.send();
-        } catch (MessagingException e) {
-            e.printStackTrace();
-        } catch (UnsupportedEncodingException e) {
+        } catch (MessagingException | UnsupportedEncodingException e) {
             e.printStackTrace();
         }
 
         return authKey;
+    }
+
+    @Transactional
+    public User signup(UserDto userDto){
+        // 중복 처리
+        if(userRepository.findByUserEmail(userDto.getUserEmail()).orElse(null) != null){
+            throw new RuntimeException("이미 가입된 유저입니다.");
+        }
+
+        User user = User.builder()
+                .email(userDto.getUserEmail())
+                .password(passwordEncoder.encode(userDto.getPassword()))
+                .name(userDto.getUserName())
+                .build();
+
+        return userRepository.save(user);
+    }
+
+    @Transactional
+    public void updateAuthStatus(String email) {
+        User user = userRepository.findByUserEmail(email).orElseThrow();
+        user.setAuthStatus(true);
     }
 
 }
